lambdas/generator/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.2
distro==1.8.0
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.0.64
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.0
packaging==23.2
playwright==1.39.0
pydantic==2.5.1
pydantic_core==2.14.3
pyee==11.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
